pyproject.toml
[tool.poetry]
name = "aieval-backend"
version = "0.1.0"
description = ""
authors = ["Charan A A <mariswarycharan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.7.1"
google-ai-generativelanguage = "^0.6.6"
fastapi = {extras = ["standard"], version = "^0.112.2"}
uvicorn = "*"
python-dotenv = "*"
pydantic = "*"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
